BJS2-44396 refactored folders names
