+++ src/main/java/school/faang/sprint1/taskbjs243512/HogwartsSpells.java
@@ -1,4 +1,4 @@
-package school.faang.sprint_1.task_BJS2_43512;
+package school.faang.sprint1.taskbjs243512;
 
 import java.util.ArrayList;
 import java.util.HashMap;
