+++ src/main/java/school/faang/sprint1/taskbjs243512/SpellEvent.java
@@ -1,4 +1,4 @@
-package school.faang.sprint_1.task_BJS2_43512;
+package school.faang.sprint1.taskbjs243512;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
